flexbox set up on current weather bar
