+++ src/App.css
@@ -30,10 +30,10 @@
 .card-container{
   display: flex;
   flex-direction: row;
+  flex-wrap: wrap;
   justify-content: center;
   align-items: center;
   gap: 2vw;
-  padding-top: 5vh;
 }
 
 .card {
@@ -47,10 +47,10 @@
   font-weight: bold;
 }
 
-.current-card {
-  flex-grow: 2;
-  /* width: 100vw; */
-}
+/* .current-card { */
+  /* flex-grow: 2;
+  width: 100vw;
+} */
 
 .locationBanner {
   width: 100vw;
